refactor: #13 - MatchIdConvert, ChampionStatsConvert, MatchMapper 클래스들을 사용되는 도메인의 util 패키지로 위치 이전
- ChampionStat 클래스를 entity/model 패키지로 위치 이전
