+++ backend/ezgg/src/main/java/com/matching/ezgg/data/memberInfo/util/MatchIdConvert.java
@@ -1,4 +1,4 @@
-package com.matching.ezgg.global.common;
+package com.matching.ezgg.data.memberInfo.util;
 
 import java.io.IOException;
 import java.util.List;
