+++ backend/ezgg/src/main/java/com/matching/ezgg/data/recentTwentyMatch/dto/RecentTwentyMatchDto.java
@@ -3,7 +3,7 @@ package com.matching.ezgg.data.recentTwentyMatch.dto;
 import java.util.Map;
 
 import com.matching.ezgg.data.recentTwentyMatch.entity.RecentTwentyMatch;
-import com.matching.ezgg.data.recentTwentyMatch.ChampionStat;
+import com.matching.ezgg.data.recentTwentyMatch.entity.model.ChampionStat;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
